agrego id a patos

+++ server/game_loop.h
@@ -29,8 +29,8 @@ private:
     ProtectedQueuesMap& queues_map;
     std::map<uint8_t, DuckPlayer> map_personajes;
     std::map<uint8_t,Weapon> map_free_weapons;
-    std::map<uint8_t, std::unique_ptr<Bullet>> map_bullets;
-    uint8_t id_balas;
+    std::map<uint16_t, std::unique_ptr<Bullet>> map_bullets;
+    uint16_t id_balas;
     void paraCadaPatoAction();
 
 public:
